- Updated podspec - Updated readme to v2.8.0 - Removed static lib file - Added JWPlayer_iOS_SDK.framework - Removed headers folder
